Re arrange app configurations, tidy up middleswares, helment, views, routes etc...
